# backend/artifacts/openapi-v6.0.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "ClearPath Nexus Engine",
    "version": "6.0.0"
  },
  "paths": {
...
    "/api/v1/multimodal/plans": {
      "post": {
        "tags": [
          "multimodal-planning"
        ],
        "summary": "Create Multimodal Plan",
        "description": "Evaluate operator/provider-supplied legs without inventing missing values.",
        "operationId": "create_multimodal_plan_api_v1_multimodal_plans_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MultimodalPlanCreate"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MultimodalPlanResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "multimodal-planning"
        ],
        "summary": "List Multimodal Plans",
        "operationId": "list_multimodal_plans_api_v1_multimodal_plans_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "default": 25,
              "title": "Limit"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/MultimodalPlanResponse"
                  },
                  "title": "Response List Multimodal Plans Api V1 Multimodal Plans Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/multimodal/plans/{plan_id}": {
      "get": {
        "tags": [
          "multimodal-planning"
        ],
        "summary": "Read Multimodal Plan",
        "operationId": "read_multimodal_plan_api_v1_multimodal_plans__plan_id__get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "plan_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Plan Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MultimodalPlanResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/operations/overview": {
      "get": {
        "tags": [
          "integrated-operations"
        ],
        "summary": "Operations Overview",
        "operationId": "operations_overview_api_v1_operations_overview_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OperationsOverview"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/v1/operations/twin": {
      "get": {
        "tags": [
          "integrated-operations"
        ],
        "summary": "Operations Twin",
        "operationId": "operations_twin_api_v1_operations_twin_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 250,
              "minimum": 1,
              "default": 100,
              "title": "Limit"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OperationsTwin"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/operations/audit/export": {
      "get": {
        "tags": [
          "integrated-operations"
        ],
        "summary": "Audit Export",
        "operationId": "audit_export_api_v1_operations_audit_export_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 1000,
              "minimum": 1,
              "default": 200,
              "title": "Limit"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuditExport"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "AuditExport": {
        "properties": {
          "generated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Generated At"
          },
          "format_version": {
            "type": "string",
            "title": "Format Version"
          },
          "owner_scope": {
            "type": "string",
            "title": "Owner Scope"
          },
          "records": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "type": "array",
            "title": "Records"
          },
          "limitations": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Limitations"
          }
        },
        "type": "object",
        "required": [
          "generated_at",
          "format_version",
          "owner_scope",
          "records",
          "limitations"
        ],
        "title": "AuditExport"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "MultimodalLegInput": {
        "properties": {
          "mode": {
            "type": "string",
            "enum": [
              "ROAD",
              "RAIL",
              "PORT",
              "SEA"
            ],
            "title": "Mode"
          },
          "origin": {
            "type": "string",
            "maxLength": 160,
            "minLength": 2,
            "title": "Origin"
          },
          "destination": {
            "type": "string",
            "maxLength": 160,
            "minLength": 2,
            "title": "Destination"
          },
          "distance_km": {
            "type": "number",
            "maximum": 50000.0,
            "minimum": 0.0,
            "title": "Distance Km"
          },
          "estimated_minutes": {
            "type": "integer",
            "maximum": 525600.0,
            "minimum": 0.0,
            "title": "Estimated Minutes"
          },
          "cost_amount": {
            "anyOf": [
              {
                "type": "number",
                "maximum": 1000000000.0,
                "minimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Cost Amount"
          },
          "cost_source_type": {
            "type": "string",
            "enum": [
              "VERIFIED_TARIFF",
              "OPERATOR_INPUT",
              "UNAVAILABLE"
            ],
            "title": "Cost Source Type",
            "default": "UNAVAILABLE"
          },
          "risk_score": {
            "anyOf": [
              {
                "type": "number",
                "maximum": 100.0,
                "minimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Risk Score"
          },
          "risk_source_type": {
            "type": "string",
            "enum": [
              "LIVE_PROVIDER",
              "PUBLIC_OPEN_DATA",
              "CACHED_PROVIDER",
              "OPERATOR_INPUT",
              "SEEDED_BASELINE",
              "DERIVED",
              "SIMULATED",
              "UNAVAILABLE"
            ],
            "title": "Risk Source Type",
            "default": "UNAVAILABLE"
          },
          "compliance_status": {
            "type": "string",
            "enum": [
              "PASSED",
              "WARNING",
              "BLOCKED",
              "MANUAL_REVIEW",
              "NOT_ASSESSED"
            ],
            "title": "Compliance Status",
            "default": "NOT_ASSESSED"
          },
          "compliance_check_id": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Compliance Check Id"
          },
          "source_provider": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 120
              },
              {
                "type": "null"
              }
            ],
            "title": "Source Provider"
          },
          "source_dataset": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 160
              },
              {
                "type": "null"
              }
            ],
            "title": "Source Dataset"
          },
          "source_reference": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 500
              },
              {
                "type": "null"
              }
            ],
            "title": "Source Reference"
          },
          "source_observed_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Source Observed At"
          },
          "constraints": {
            "items": {
              "$ref": "#/components/schemas/LegConstraint"
            },
            "type": "array",
            "maxItems": 30,
            "title": "Constraints"
          }
        },
        "type": "object",
        "required": [
          "mode",
          "origin",
          "destination",
          "distance_km",
          "estimated_minutes"
        ],
        "title": "MultimodalLegInput"
      },
...
      "MultimodalPlanCreate": {
        "properties": {
          "name": {
            "type": "string",
            "maxLength": 160,
            "minLength": 2,
            "title": "Name"
          },
          "origin": {
            "type": "string",
            "maxLength": 160,
            "minLength": 2,
            "title": "Origin"
          },
          "destination": {
            "type": "string",
            "maxLength": 160,
            "minLength": 2,
            "title": "Destination"
          },
          "cost_currency": {
            "type": "string",
            "pattern": "^[A-Z]{3}$",
            "title": "Cost Currency",
            "default": "INR"
          },
          "legs": {
            "items": {
              "$ref": "#/components/schemas/MultimodalLegInput"
            },
            "type": "array",
            "maxItems": 20,
            "minItems": 1,
            "title": "Legs"
          }
        },
        "type": "object",
        "required": [
          "name",
          "origin",
          "destination",
          "legs"
        ],
        "title": "MultimodalPlanCreate"
      },
      "MultimodalPlanResponse": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "origin": {
            "type": "string",
            "title": "Origin"
          },
          "destination": {
            "type": "string",
            "title": "Destination"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "recommendation": {
            "type": "string",
            "title": "Recommendation"
          },
          "total_distance_km": {
            "type": "number",
            "title": "Total Distance Km"
          },
          "total_eta_minutes": {
            "type": "integer",
            "title": "Total Eta Minutes"
          },
          "total_cost": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Total Cost"
          },
          "cost_currency": {
            "type": "string",
            "title": "Cost Currency"
          },
          "cost_completeness": {
            "type": "number",
            "title": "Cost Completeness"
          },
          "overall_risk_score": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Overall Risk Score"
          },
          "compliance_status": {
            "type": "string",
            "title": "Compliance Status"
          },
          "traceability_summary": {
            "additionalProperties": true,
            "type": "object",
            "title": "Traceability Summary"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "legs": {
            "items": {
              "$ref": "#/components/schemas/MultimodalLegResponse"
            },
            "type": "array",
            "title": "Legs"
          }
        },
        "type": "object",
        "required": [
          "id",
          "name",
          "origin",
          "destination",
          "status",
          "recommendation",
          "total_distance_km",
          "total_eta_minutes",
          "total_cost",
          "cost_currency",
          "cost_completeness",
          "overall_risk_score",
          "compliance_status",
          "traceability_summary",
          "created_at",
          "legs"
        ],
        "title": "MultimodalPlanResponse"
      },
      "OperationsOverview": {
        "properties": {
          "generated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Generated At"
          },
          "product_version": {
            "type": "string",
            "title": "Product Version"
          },
          "routes_total": {
            "type": "integer",
            "title": "Routes Total"
          },
          "shipments_by_status": {
            "additionalProperties": {
              "type": "integer"
            },
            "type": "object",
            "title": "Shipments By Status"
          },
          "schedules_by_status": {
            "additionalProperties": {
              "type": "integer"
            },
            "type": "object",
            "title": "Schedules By Status"
          },
          "open_events_by_severity": {
            "additionalProperties": {
              "type": "integer"
            },
            "type": "object",
            "title": "Open Events By Severity"
          },
          "compliance_by_status": {
            "additionalProperties": {
              "type": "integer"
            },
            "type": "object",
            "title": "Compliance By Status"
          },
          "multimodal_by_status": {
            "additionalProperties": {
              "type": "integer"
            },
            "type": "object",
            "title": "Multimodal By Status"
          },
          "train_sync_by_status": {
            "additionalProperties": {
              "type": "integer"
            },
            "type": "object",
            "title": "Train Sync By Status"
          },
          "prediction_count": {
            "type": "integer",
            "title": "Prediction Count"
          },
          "traceability": {
            "additionalProperties": true,
            "type": "object",
            "title": "Traceability"
          },
          "provider_health": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "type": "array",
            "title": "Provider Health"
          },
          "notices": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Notices"
          }
        },
        "type": "object",
        "required": [
          "generated_at",
          "product_version",
          "routes_total",
          "shipments_by_status",
          "schedules_by_status",
          "open_events_by_severity",
          "compliance_by_status",
          "multimodal_by_status",
          "train_sync_by_status",
          "prediction_count",
          "traceability",
          "provider_health",
          "notices"
        ],
        "title": "OperationsOverview"
      },
      "OperationsTwin": {
        "properties": {
          "generated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Generated At"
          },
          "model_type": {
            "type": "string",
            "title": "Model Type"
          },
          "shipments": {
            "items": {
              "$ref": "#/components/schemas/ShipmentTwin"
            },
            "type": "array",
            "title": "Shipments"
          },
          "disclaimer": {
            "type": "string",
            "title": "Disclaimer"
          }
        },
        "type": "object",
        "required": [
          "generated_at",
          "model_type",
          "shipments",
          "disclaimer"
        ],
        "title": "OperationsTwin"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}